Fix CharScript names & Add Collision Events
