+++ Content/LeagueSandbox-Scripts/Characters/Sion/Passive.cs
@@ -8,35 +8,41 @@ using LeagueSandbox.GameServer.API;
 using GameServerCore.Domain;
 using GameServerLib.GameObjects.AttackableUnits;
 using static LeagueSandbox.GameServer.API.ApiFunctionManager;
+using GameServerCore.Enums;
 
-
-namespace CharScripts
+namespace Spells
 {
-    public class CharScriptSion : ICharScript
+    public class SionPassiveSpeed : ISpellScript
     {
-        ISpell Spell;
-        int counter;
-        public void OnActivate(IObjAiBase owner, ISpell spell = null)
+        public ISpellScriptMetadata ScriptMetadata => new SpellScriptMetadata()
         {
-            ApiEventManager.OnDeath.AddListener(this, owner, OnDeath, true);
-            ApiEventManager.OnResurrect.AddListener(this, owner, OnRessurect, false);
-            Spell = spell;
+            TriggersSpellCasts = true
+            // TODO
+        };
+
+        public void OnActivate(IObjAiBase owner, ISpell spell)
+        {
         }
-        public void OnDeath(IDeathData deathData)
+        public void OnDeactivate(IObjAiBase owner, ISpell spell)
         {
-            AddBuff("SionPassiveDelay", 2f, 1, Spell, deathData.Unit, deathData.Unit as IObjAiBase);
-        } 
-        public void OnRessurect(IObjAiBase owner)
-        {
-            counter++;
-            //This is to avoid a loop in his passive.
-            if (counter == 2)
-            {
-                ApiEventManager.OnDeath.AddListener(this, owner, OnDeath, true); 
-                counter = 0;
-            }
         }
-        public void OnDeactivate(IObjAiBase owner, ISpell spell = null)
+        public void OnSpellPreCast(IObjAiBase owner, ISpell spell, IAttackableUnit target, Vector2 start, Vector2 end)
+        {
+        }
+        public void OnSpellCast(ISpell spell)
+        {
+            AddBuff("SionPassiveSpeed", 1.5f, 1, spell, spell.CastInfo.Owner, spell.CastInfo.Owner);
+        }
+        public void OnSpellPostCast(ISpell spell)
+        {
+        }
+        public void OnSpellChannel(ISpell spell)
+        {
+        }
+        public void OnSpellChannelCancel(ISpell spell, ChannelingStopSource reason)
+        {
+        }
+        public void OnSpellPostChannel(ISpell spell)
         {
         }
         public void OnUpdate(float diff)
